+++ packages/spotlight/src/spotlight.js
@@ -162,6 +162,10 @@ const Spotlight = (function () {
 
 		let currentFocusedElement = getCurrent();
 
+		if (elem === currentFocusedElement) {
+			return true;
+		}
+
 		let silentFocus = function () {
 			if (currentFocusedElement) {
 				currentFocusedElement.blur();